requirements.txt
Flask==2.0.2
python-dotenv==0.19.1
SQLAlchemy==1.4.27
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
email-validator==1.1.3
bcrypt==3.2.0
python-jose==3.3.0
requests==2.26.0
openai==0.10.2
alembic==1.7.5
psycopg2-binary==2.9.1
gunicorn==20.1.0
celery==5.1.2
redis==4.0.2
flower==1.0.0
pandas==1.3.4
matplotlib==3.5.0
numpy==1.21.4
scikit-learn==1.0.1
react==17.0.2
react-dom==17.0.2
react-router-dom==5.3.0
react-scripts==4.0.3
axios==0.24.0
bootstrap==5.1.3
reactstrap==8.10.0
jwt-decode==3.1.2
formik==2.2.9
yup==0.32.11
chart.js==3.6.0
react-chartjs-2==3.0.5
react-bootstrap==2.0.2
bootstrap-icons==1.7.2